requirements.txt
langchain
langchain_anthropic
python-dotenv
langchain_community
pymupdf
fastapi
uvicorn